Refactor struct tests
