--- modules/ballerina-test/src/test/resources/test-src/structs/duplicate-fields.bal
@@ -1,6 +1,0 @@
-struct Department {
-    string dptName;
-    int id;
-    string id;
-    Person prsn1;
-}
