--- modules/ballerina-test/src/test/resources/test-src/structs/invalid-field-name-init.bal
@@ -1,9 +1,0 @@
-function testInvalidTypeAttributeinit () {
-    string name;
-    Department dpt = {dptName[0]:54};
-}
-
-struct Department {
-    string[] dptName;
-    int count;
-}
