--- modules/ballerina-test/src/test/resources/test-src/structs/invalid-type-attribute-init.bal
@@ -1,9 +1,0 @@
-function testInvalidTypeAttributeinit () {
-    string name;
-    Department dpt = {dptNames:54};
-}
-
-struct Department {
-    map dptNames;
-    int count;
-}
