+++ modules/ballerina-test/src/test/resources/test-src/structs/struct-access-with-index-negative.bal
@@ -1,0 +1,15 @@
+function testUndeclaredStructAccess() {        
+    string name;        
+    dpt1[name] = "HR";      
+}
+
+function testUndeclaredAttributeAccess() {     
+    string name;        
+    Department dpt;     
+    dpt["id"] = "HR";       
+}     
+      
+struct Department {       
+    string dptName;     
+    int count;      
+}
