+++ modules/ballerina-test/src/test/resources/test-src/structs/struct-access-with-index.bal
