--- modules/ballerina-test/src/test/resources/test-src/structs/undeclared-attribute-init.bal
@@ -1,9 +1,0 @@
-function testUndeclaredAttributeinit () {
-    string name;
-    Department dpt = {dptName:"HR", age:20};
-}
-
-struct Department {
-    string dptName;
-    int count;
-}
