--- modules/ballerina-test/src/test/resources/test-src/structs/undeclared-struct-access.bal
@@ -1,5 +1,0 @@
-function testUndeclaredStructAccess () {
-    string name;
-
-    dpt1.name = "HR";
-}
