--- modules/ballerina-test/src/test/resources/test-src/structs/undeclared-struct-init.bal
@@ -1,3 +1,0 @@
-function testUndeclaredStructInit () {
-    Department dpt1 = {name:"HR"};
-}
