+++ modules/ballerina-test/src/test/resources/testng.xml
@@ -29,6 +29,7 @@ under the License.
             <package name="org.ballerinalang.test.types.*"/>
             <package name="org.ballerinalang.test.nativeImpl.*"/>
             <package name="org.ballerinalang.test.net.*"/>
+            <package name="org.ballerinalang.test.structs.*"/>
         </packages>
     </test>
 </suite>
